requirements.txt
Django>=5.2.6
psycopg2-binary>=2.9.9
Pillow>=10.0.0
python-decouple>=3.8
django-allauth>=0.57.0